Added +layout.svelte and created utils.ts to store frontend utils
