+++ src/routes/+layout.svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+  import { initTauriCore } from "$lib/utils";
+  import { onMount } from "svelte";
+
+  onMount(() => {
+    initTauriCore();
+  });
+
+  const { children } = $props();
+</script>
+
+{@render children()}
